+++ include/boost/signals2/variadic_signal.hpp
@@ -16,13 +16,13 @@
 #ifndef BOOST_SIGNALS2_VARIADIC_SIGNAL_HPP
 #define BOOST_SIGNALS2_VARIADIC_SIGNAL_HPP
 
+#include <boost/core/enable_if.hpp>
 #include <boost/preprocessor/control/expr_if.hpp>
 #include <boost/signals2/detail/variadic_arg_type.hpp>
 #include <boost/signals2/detail/variadic_slot_invoker.hpp>
 #include <boost/smart_ptr/make_shared.hpp>
 #include <boost/type_traits/function_traits.hpp>
 #include <boost/type_traits/is_void.hpp> 
-#include <boost/utility/enable_if.hpp>
 
 namespace boost
 {
